requirements.txt
numpy
pandas
openai
python-dotenv
fastapi 
uvicorn